refact method water: parameter type is rewritten from int to float

+++ week-10/day-2/GardenApplication2/src/Plant.java
@@ -18,7 +18,7 @@ public class Plant {
         return waterLevel < criticalWaterLevel;
     }
 
-    public void water(int waterAmount) {
+    public void water(float waterAmount) {
         waterLevel += waterAmount * waterAbsorption;
     }
